# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PROJECT HELIOS - AI Solar Flare Forecasting API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FluxReading {
  id: number;
  timeTag: string;
  /** @nullable */
  softXray?: number | null;
  /** @nullable */
  hardXray?: number | null;
  channel: string;
  /** @nullable */
  satellite?: number | null;
  /** @nullable */
  flareClass?: string | null;
}

export interface SolarStatus {
  currentFlux: number;
  flareClass: string;
  flareClassLetter: string;
  riskLevel: string;
  softXray: number;
  hardXray: number;
  lastUpdated: string;
  systemStatus: string;
}

export interface FlareEvent {
  id: number;
  timeTag: string;
  flareClass: string;
  peakFlux: number;
  confidence: number;
  source: string;
  /** @nullable */
  duration?: number | null;
  status: string;
}

export interface Nowcast {
  detectedClass: string;
  confidence: number;
  peakFlux: number;
  currentEnergy: number;
  source: string;
  timestamp: string;
  isActive: boolean;
  riskLevel: string;
  explanation: string;
}

export interface SolarSummary {
  totalEventsDetected: number;
  predictionAccuracy: number;
  currentSolarActivity: string;
  averageLeadTime: number;
  activeAlerts: number;
  systemStatus: string;
  modelStatus: string;
  databaseStatus: string;
  todayEventCount: number;
  weeklyTrend: number;
}

export interface ForecastResult {
  horizonMinutes: number;
  probability: number;
  confidence: number;
  expectedClass: string;
  estimatedPeak: string;
  leadTime: number;
  trend: string;
}

export interface Alert {
  id: number;
  type: string;
  severity: string;
  title: string;
  description: string;
  /** @nullable */
  flareClass?: string | null;
  /** @nullable */
  confidence?: number | null;
  timestamp: string;
  isActive: boolean;
  /** @nullable */
  countdown?: number | null;
}

export interface DailyAnalytics {
  date: string;
  eventCount: number;
  maxFlux: number;
  avgFlux: number;
  xClassCount: number;
  mClassCount: number;
  cClassCount: number;
  bClassCount: number;
}

export interface HeatmapEntry {
  hour: number;
  day: number;
  value: number;
  label: string;
}

export interface ClassCount {
  flareClass: string;
  count: number;
  percentage: number;
  color: string;
}

export interface ModelPerformance {
  accuracy: number;
  precision: number;
  recall: number;
  f1Score: number;
  auc: number;
  falseAlarmRate: number;
  missRate: number;
  trainingDataSize: number;
  lastTrainedAt: string;
  modelName: string;
  version: string;
  confusionMatrix: number[][];
}

export interface FeatureImportance {
  feature: string;
  importance: number;
  category: string;
}

export interface Dataset {
  id: number;
  name: string;
  type: string;
  recordCount: number;
  uploadedAt: string;
  status: string;
  /** @nullable */
  channelInfo?: string | null;
}

export type GetSolarFluxParams = {
channel?: GetSolarFluxChannel;
limit?: number;
};

export type GetSolarFluxChannel = typeof GetSolarFluxChannel[keyof typeof GetSolarFluxChannel];


export const GetSolarFluxChannel = {
  soft: 'soft',
  hard: 'hard',
  both: 'both',
} as const;

export type GetSolarEventsParams = {
limit?: number;
};

export type GetAlertsParams = {
limit?: number;
};

export type GetDailyAnalyticsParams = {
days?: number;
};