# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PROJECT HELIOS - AI Solar Flare Forecasting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /solar/flux:
    get:
      operationId: getSolarFlux
      tags: [solar]
      summary: Get recent solar X-ray flux readings
      parameters:
        - name: channel
          in: query
          schema:
            type: string
            enum: ["soft", "hard", "both"]
            default: "both"
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Solar flux readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FluxReading"
...
  /solar/events:
    get:
      operationId: getSolarEvents
      tags: [solar]
      summary: Get detected solar flare events
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Solar flare events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlareEvent"
...
components:
  schemas:
...
    FluxReading:
      type: object
      properties:
        id:
          type: integer
        timeTag:
          type: string
        softXray:
          type: ["number", "null"]
        hardXray:
          type: ["number", "null"]
        channel:
          type: string
        satellite:
          type: ["integer", "null"]
        flareClass:
          type: ["string", "null"]
      required:
        - id
        - timeTag
        - channel
...
    FlareEvent:
      type: object
      properties:
        id:
          type: integer
        timeTag:
          type: string
        flareClass:
          type: string
        peakFlux:
          type: number
        confidence:
          type: number
        source:
          type: string
        duration:
          type: ["integer", "null"]
        status:
          type: string
      required:
        - id
        - timeTag
        - flareClass
        - peakFlux
        - confidence
        - source
        - status
...