# lib/api-zod/src/generated/types/flareEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PROJECT HELIOS - AI Solar Flare Forecasting API
 * OpenAPI spec version: 0.1.0
 */

export interface FlareEvent {
  id: number;
  timeTag: string;
  flareClass: string;
  peakFlux: number;
  confidence: number;
  source: string;
  /** @nullable */
  duration?: number | null;
  status: string;
}